Adds Sass support, Laravel Mix as dedicated asset pipeline
The addition of Webpack (Mix) is kinda overkill, and adds a fair amount
of package.json script bloat and output complexity. Hot reloading
currently not working with npm scripts and not sure if that was a
feature of 11ty to begin with.

Layouts transitioned to using Liquid templating language, since Nunjucks
doesn't have a clear way to get syntax highlighting in Neovim.